feat(beneath): add support for "audit functions"
At each step of the path resolution, and when any symlinks are
encountered, the audit function will be called with relevant information
about the situation.

This can be used to impose extra permission checks (for example, if the
file is being opened on behalf of another user, and it's important to
make sure they have access to every component as it's resolved).